Moved FFmpeg to product flavors
