+++ ffmpeg/src/bundled/java/com/yausername/ffmpeg/FFmpeg.kt
@@ -1,0 +1,30 @@
+package com.yausername.ffmpeg
+
+import android.content.Context
+import com.yausername.youtubedl_android.util.exceptions.YoutubeDLException
+import com.yausername.youtubedl_common.Constants
+import com.yausername.youtubedl_common.utils.ZipUtils.unzip
+import org.apache.commons.io.FileUtils
+import java.io.File
+
+object FFmpeg: FFmpegCore() {
+    override fun initFFmpeg(appContext: Context, ffmpegDir: File) {
+        val ffmpegLib = File(binDir, Constants.LibrariesName.FFMPEG)
+        // using size of lib as version
+        val ffmpegSize = ffmpegLib.length().toString()
+        if (!ffmpegDir.exists() || shouldUpdateFFmpeg(appContext, ffmpegSize)) {
+            FileUtils.deleteQuietly(ffmpegDir)
+            ffmpegDir.mkdirs()
+            try {
+                unzip(ffmpegLib, ffmpegDir)
+            } catch (e: Exception) {
+                FileUtils.deleteQuietly(ffmpegDir)
+                throw YoutubeDLException("failed to initialize", e)
+            }
+            updateFFmpeg(appContext, ffmpegSize)
+        }
+    }
+
+    @JvmStatic
+    fun getInstance() = this
+}
